backend/requirements.txt
# 企业级RAG系统 - 后端依赖

# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
alembic==1.13.1
sqlalchemy==2.0.25
sqlmodel==0.0.14
psycopg2-binary==2.9.9

# 向量数据库
pymilvus==2.3.4

# RAG核心
langchain==0.1.5
langchain-community==0.0.18
langchain-openai==0.0.2

# 文档处理
pypdf==4.0.1
docx2txt==0.8
python-docx==1.1.0
openpyxl==3.1.2

# AI模型
openai==1.10.0  # 兼容OpenAI格式的国内API
httpx==0.26.0

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 工具
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
celery==5.3.6
redis==5.0.1

# 前端
jinja2==3.1.3

# 测试
pytest==8.0.0
pytest-asyncio==0.23.3
httpx
